requirements.txt
Flask==2.3.2
Pillow==10.0.0
opencv-python-headless==4.8.0.76
numpy==1.24.4
werkzeug==2.3.7
tensorflow-cpu==2.10.0
